Fix: - Made adjustments according to the recommendations
