--- src/com/company/BloomFilter/BloomFilter.java
@@ -1,9 +1,0 @@
-package com.company.BloomFilter;
-
-/**
- * BloomFilter implementation interface
- */
-interface BloomFilter {
-    void add(String value);
-    boolean mightContain(String value);
-}
